long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5971 a1c6a512-1295-4272-9138-f99709370657

+++ apps/recorder/recording.c
@@ -241,7 +241,7 @@ int rec_create_directory(void)
 
 bool recording_screen(void)
 {
-    int button;
+    long button;
     bool done = false;
     char buf[32];
     char buf2[32];
